feature of presentation  messages during auth process according to auth actions
