--- src/client/js/helpers/renderErrorReport.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default (msg) => {
-  const stylesDescr = {
-    fontSize: '19px',
-  };
-
-  return msg ? <p style={stylesDescr}>{msg}</p> : null;
-};
